third_party/blink/web_tests/svg/dynamic-updates/SVGRadialGradientElement-dom-cy-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cy' attribute of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.getAttribute('cy') is "0%"
PASS radialGradientElement.getAttribute('cy') is "150%"
PASS successfullyParsed is true

TEST COMPLETE
